# Challenge
An attempt at the challenge for VBA.
 1. Created arrays for volumes, starting prices and ending prices
2. expanded search and index to all tickers not just the 12 initially
3. Removed nested loops as very row was being evaluated not just selected tickers
4. Added ticker Index

Regrettably, I am unable to get to run as the setup for the index is not correct.  I was unable to source any guidance from the module as this was not anything we were exposed to in class or in exercises.  I have asked for a tutor as this section was terribly defeating